basic mouse support (refactored from @gwicho38)
