+++ examples/default-config.toml
@@ -11,7 +11,7 @@
 startup_commands = []
 refresh_fps = 10
 network_timeout_seconds = 60
-enable_mouse = false
+mouse_support = false
 article_scope = "unread"
 feed_list_scope = "all"
 default_sort_order = "<date"
