+++ src/ui/view.rs
@@ -47,10 +47,7 @@ impl Widget for &mut App {
         let (articles_constraint_height, article_content_constraint_height) =
             if let Some(override_height) = self.articles_height_override {
                 // User is dragging the border — use absolute heights
-                (
-                    Constraint::Length(override_height),
-                    Constraint::Min(0),
-                )
+                (Constraint::Length(override_height), Constraint::Min(0))
             } else {
                 match self.state {
                     AppState::FeedSelection | AppState::ArticleSelection => (
@@ -82,9 +79,9 @@ impl Widget for &mut App {
             .areas(articles_chunk);
 
         // store areas for mouse hit-testing
-        self.panel_areas.feed_list = feeds_list_chunk;
-        self.panel_areas.articles_list = articles_list_chunk;
-        self.panel_areas.article_content = article_content_chunk;
+        *self.panel_areas.feed_list_mut() = feeds_list_chunk;
+        *self.panel_areas.articles_list_mut() = articles_list_chunk;
+        *self.panel_areas.article_content_mut() = article_content_chunk;
 
         // render stuff
         self.feed_list.render(feeds_list_chunk, buf);
